fix: refactored and clean everything to follow one smooth process
