+++ ecs/web/examples/rush-keystore-nextjs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test",
+  "name": "rush-keystore-nextjs",
   "version": "0.1.0",
   "private": true,
   "scripts": {
