+++ ecs/web/examples/rush-keystore-nextjs/src/app/favicon.ico
