+++ ecs/web/sdk/typescript/src/tests/test.ts
@@ -1,0 +1,1 @@
+// Write Test here. Removed since restarting
