+++ ecs/web/sdk/typescript/tsconfig.json
@@ -13,7 +13,7 @@
 		"declaration": true,
 		"outDir": "./dist",
 		"rootDir": "./src",
-		"types": ["./src/types/types.d.ts", "node", "jest"] // Added jest to the types field
+		"types": ["./src/core/types/types.d.ts", "node", "jest"] // Added jest to the types field
 	},
 	"include": ["*/.d.ts", "src/**/*.ts"],
 	"exclude": ["node_modules"]
